# .agents/skills/hashtagweb3-agent-skill/public/openapi.yaml
openapi: 3.1.0
info:
  title: Hashtag Web3 Public API
  version: 1.0.0
  description: Public REST API for Hashtag Web3 (hashtagweb3.com). Search live Web3 jobs, crypto industry news, global blockchain conferences & hackathons, and the 200+ term Web3 glossary. Optimized for LLM agents, automated scripts, and developer integrations.
  x-api-versioning:
    policy: "URI path versioning (/api/v1/)"
    deprecation: "We signal deprecation using standard Deprecation and Sunset HTTP headers in API responses, with a minimum 12-month advance notice period."
    policyUrl: https://hashtagweb3.com/api-policy
    documentationUrl: https://hashtagweb3.com/developers
  contact:
    name: Hashtag Web3 Developer Support
    url: https://hashtagweb3.com/developers
    email: contact@hashtagweb3.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://hashtagweb3.com/api/v1
    description: Production server (v1)
  - url: https://hashtagweb3.com/api/sandbox
    description: Sandbox / Test Environment (Mock Data, Zero-Auth, Safe Testing)
  - url: https://hashtagweb3.com
    description: Production server root
paths:
...
  /api/glossary:
    get:
      operationId: listGlossaryTerms
      summary: Search Web3 & Blockchain Glossary
      description: Query 200+ technical Web3 glossary terms across consensus, DeFi, cryptography, Layer 2 scaling, and tokenomics.
      parameters:
        - name: search
          in: query
          required: false
          description: Search term name, definition, or synonyms.
          schema:
            type: string
            example: Zero Knowledge
        - name: category
          in: query
          required: false
          description: Filter by glossary category.
          schema:
            type: string
            example: DeFi
        - name: limit
          in: query
          required: false
          description: Maximum number of terms to return.
          schema:
            type: integer
            default: 50
            example: 20
      responses:
        '200':
          description: List of matching glossary terms.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlossaryListResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  schemas:
...
    GlossaryTerm:
      type: object
      required: [slug, term, category, definition]
      properties:
        slug:
          type: string
          example: zero-knowledge-proofs
        term:
          type: string
          example: Zero-Knowledge Proofs (ZKP)
        category:
          type: string
          example: Cryptography
        definition:
          type: string
          example: A cryptographic method by which one party can prove to another that a given statement is true without conveying any information apart from the fact that the statement is indeed true.
        url:
          type: string
          format: uri
          example: https://hashtagweb3.com/zero-knowledge-proofs
    GlossaryListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GlossaryTerm'
        meta:
          type: object
          required: [total, count]
          properties:
            total:
              type: integer
              example: 210
            count:
              type: integer
              example: 50
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, hint]
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, NOT_FOUND, INTERNAL_SERVER_ERROR, RATE_LIMITED]
              example: BAD_REQUEST
            message:
              type: string
              example: "Invalid parameter: 'limit' must be an integer between 1 and 200."
            hint:
              type: string
              example: "Adjust the limit parameter and retry. View docs at https://hashtagweb3.com/developers"
            docUrl:
              type: string
              format: uri
              example: https://hashtagweb3.com/developers